Separate formatter from textUi.
