+++ src/seedu/addressbook/ui/TextUi.java
@@ -23,16 +23,16 @@ public class TextUi {
 
     /** A decorative prefix added to the beginning of lines printed by AddressBook */
     private static final String LINE_PREFIX = "|| ";
-
-    /** A platform independent line separator. */
-    private static final String LS = System.lineSeparator();
-
-    private static final String DIVIDER = "===================================================";
-
-    /** Format of indexed list item */
-    private static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
-
-
+//
+//    /** A platform independent line separator. */
+//    private static final String LS = System.lineSeparator();
+//
+//    private static final String DIVIDER = "===================================================";
+//
+//    /** Format of indexed list item */
+//    private static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
+//
+//
     /** Offset required to convert between 1-indexing and 0-indexing.  */
     public static final int DISPLAYED_INDEX_OFFSET = 1;
 
@@ -41,15 +41,18 @@ public class TextUi {
 
     private final Scanner in;
     private final PrintStream out;
+    private Formatter fmt = null;
 
     public TextUi() {
-        this(System.in, System.out);
+        this(System.in, System.out, new Formatter());
     }
 
-    public TextUi(InputStream in, PrintStream out) {
+    public TextUi(InputStream in, PrintStream out, Formatter fmt) {
         this.in = new Scanner(in);
         this.out = out;
+        this.fmt = fmt;
     }
+
 
     /**
      * Returns true if the user input line should be ignored.
@@ -61,6 +64,7 @@ public class TextUi {
     private boolean shouldIgnore(String rawInputLine) {
         return rawInputLine.trim().isEmpty() || isCommentLine(rawInputLine);
     }
+
 
     /**
      * Returns true if the user input line is a comment line.
@@ -87,86 +91,86 @@ public class TextUi {
             fullInputLine = in.nextLine();
         }
 
-        showToUser("[Command entered:" + fullInputLine + "]");
+        fmt.showToUser("[Command entered:" + fullInputLine + "]");
         return fullInputLine;
     }
+//
+//
+//    public void showWelcomeMessage(String version, String storageFilePath) {
+//        String storageFileInfo = String.format(MESSAGE_USING_STORAGE_FILE, storageFilePath);
+//        showToUser(
+//                DIVIDER,
+//                DIVIDER,
+//                MESSAGE_WELCOME,
+//                version,
+//                MESSAGE_PROGRAM_LAUNCH_ARGS_USAGE,
+//                storageFileInfo,
+//                DIVIDER);
+//    }
 
-
-    public void showWelcomeMessage(String version, String storageFilePath) {
-        String storageFileInfo = String.format(MESSAGE_USING_STORAGE_FILE, storageFilePath);
-        showToUser(
-                DIVIDER,
-                DIVIDER,
-                MESSAGE_WELCOME,
-                version,
-                MESSAGE_PROGRAM_LAUNCH_ARGS_USAGE,
-                storageFileInfo,
-                DIVIDER);
-    }
-
-    public void showGoodbyeMessage() {
-        showToUser(MESSAGE_GOODBYE, DIVIDER, DIVIDER);
-    }
-
-
-    public void showInitFailedMessage() {
-        showToUser(MESSAGE_INIT_FAILED, DIVIDER, DIVIDER);
-    }
-
-    /** Shows message(s) to the user */
-    public void showToUser(String... message) {
-        for (String m : message) {
-            out.println(LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX));
-        }
-    }
-
-    /**
-     * Shows the result of a command execution to the user. Includes additional formatting to demarcate different
-     * command execution segments.
-     */
-    public void showResultToUser(CommandResult result) {
-        final Optional<List<? extends ReadOnlyPerson>> resultPersons = result.getRelevantPersons();
-        if (resultPersons.isPresent()) {
-            showPersonListView(resultPersons.get());
-        }
-        showToUser(result.feedbackToUser, DIVIDER);
-    }
-
-    /**
-     * Shows a list of persons to the user, formatted as an indexed list.
-     * Private contact details are hidden.
-     */
-    private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
-        final List<String> formattedPersons = new ArrayList<>();
-        for (ReadOnlyPerson person : persons) {
-            formattedPersons.add(person.getAsTextHidePrivate());
-        }
-        showToUserAsIndexedList(formattedPersons);
-    }
-
-    /** Shows a list of strings to the user, formatted as an indexed list. */
-    private void showToUserAsIndexedList(List<String> list) {
-        showToUser(getIndexedListForViewing(list));
-    }
-
-    /** Formats a list of strings as a viewable indexed list. */
-    private static String getIndexedListForViewing(List<String> listItems) {
-        final StringBuilder formatted = new StringBuilder();
-        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
-        for (String listItem : listItems) {
-            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
-            displayIndex++;
-        }
-        return formatted.toString();
-    }
-
-    /**
-     * Formats a string as a viewable indexed list item.
-     *
-     * @param visibleIndex visible index for this listing
-     */
-    private static String getIndexedListItem(int visibleIndex, String listItem) {
-        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
-    }
+//    public void showGoodbyeMessage() {
+//        showToUser(MESSAGE_GOODBYE, DIVIDER, DIVIDER);
+//    }
+//
+//
+//    public void showInitFailedMessage() {
+//        showToUser(MESSAGE_INIT_FAILED, DIVIDER, DIVIDER);
+//    }
+//
+//    /** Shows message(s) to the user */
+//    public void showToUser(String... message) {
+//        for (String m : message) {
+//            out.println(LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX));
+//        }
+//    }
+//
+//    /**
+//     * Shows the result of a command execution to the user. Includes additional formatting to demarcate different
+//     * command execution segments.
+//     */
+//    public void showResultToUser(CommandResult result) {
+//        final Optional<List<? extends ReadOnlyPerson>> resultPersons = result.getRelevantPersons();
+//        if (resultPersons.isPresent()) {
+//            showPersonListView(resultPersons.get());
+//        }
+//        showToUser(result.getFeedbackToUser(), DIVIDER);
+//    }
+//
+//    /**
+//     * Shows a list of persons to the user, formatted as an indexed list.
+//     * Private contact details are hidden.
+//     */
+//    private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
+//        final List<String> formattedPersons = new ArrayList<>();
+//        for (ReadOnlyPerson person : persons) {
+//            formattedPersons.add(person.getAsTextHidePrivate());
+//        }
+//        showToUserAsIndexedList(formattedPersons);
+//    }
+//
+//    /** Shows a list of strings to the user, formatted as an indexed list. */
+//    private void showToUserAsIndexedList(List<String> list) {
+//        showToUser(getIndexedListForViewing(list));
+//    }
+//
+//    /** Formats a list of strings as a viewable indexed list. */
+//    private static String getIndexedListForViewing(List<String> listItems) {
+//        final StringBuilder formatted = new StringBuilder();
+//        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
+//        for (String listItem : listItems) {
+//            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+//            displayIndex++;
+//        }
+//        return formatted.toString();
+//    }
+//
+//    /**
+//     * Formats a string as a viewable indexed list item.
+//     *
+//     * @param visibleIndex visible index for this listing
+//     */
+//    private static String getIndexedListItem(int visibleIndex, String listItem) {
+//        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+//    }
 
 }
